Fix filly flip issues
The issue stems from the fact that fillies x are only flipped when the chair is flipped.
This is a problem, because it only checks if the parent seat is flipped or not, and ignores the times where the chair is not flipped.
The fillies retain their flip from the previous (walking) stage, and so if the filly is facing left while the chair is facing right, they will face right